refactor(review): createKeyedRoot over mapArray + move to @kolu/surface/solid [batch group 4]
4a — createKeyedRoot becomes a thin alias over solid-js's own `mapArray`:
`createMemo(mapArray(() => [key()], factory))` + an eager `createRenderEffect`. Deletes the
hand-rolled `disposePrev` bookkeeping, the manual `untrack` fence, and the hand-written
dispose ordering (~20 lines) — mapArray keys by value-identity (the same fence) and
disposes each key's per-item owner on a change or on the caller's teardown. The eager
render effect is KEPT + newly pinned: the swap disposes the old root SYNCHRONOUSLY on a key
flip (before any read), which is load-bearing for the switch-toast abort ordering.

4b — moved createKeyedRoot + connectionScoped down to the base `@kolu/surface/solid` layer
(pure solid-generic). Dissolves the import-cycle contortion that kept them in a standalone
surface-app file; `@kolu/surface-app/solid` re-exports both for compat (drishti + kolu
imports unchanged). Test co-located in @kolu/surface.

4c — ownerless misuse now THROWS the documented contract instead of silently degrading
(lazy dispose + a leaked final root): one up-front `getOwner()` throw replaces the two
`if (getOwner())` guards. Pinned.

4d — observableHolder is backed by a private `EventEmitter` + `node:events` `once(...,
{ signal })` (max-listeners lifted, since waiters are legitimately unbounded), deleting the
hand-rolled waiter Set + abort plumbing. relayStream.test green.
